Rename tool_config and update copyright headers
